refactor: tests/integration/20.command.functions.test.cjs
refactor test file to ESM

- convert to ESM
- replace ava with vitest
- replace got with node-fetch

Related to: #5698
